// packages/topic/src/types/type.ts
export enum EnumDevice {
  pc = 'pc',
  mobile = 'mobile',
  other = 'other',
  empty = '',
}
// 开发平台环境
export enum EnumPlatformEnv {
  dev = 'dev',
  test = 'test',
  prod = 'prod',
  empty = '',
}
// 页面平台环境
export enum EnumPageEnv {
  test = 'test',
  prod = 'prod',
  empty = '',
}

export enum EmpPropTypes {
  Input = 'Input',
  /** @deprecated */
  InputText = 'InputText',
  /** @deprecated */
  I18nDynamicText = 'I18nDynamicText',
  InputTextArea = 'InputTextArea',
  InputNumber = 'InputNumber',
  Slider = 'Slider',
  Radio = 'Radio',
  RadioGroup = 'RadioGroup',
  Select = 'Select',
  Switch = 'Switch',
  TimePicker = 'TimePicker',
  DatePicker = 'DatePicker',
  RangePicker = 'RangePicker',
  ColorPicker = 'ColorPicker',
  Upload = 'Upload',
  RichText = 'RichText',
  StyleEdit = 'StyleEdit',
  Checkbox = 'Checkbox',
}